# Conflicts:
#	imgui.ini

+++ RightHumanWave.cpp
@@ -53,7 +53,7 @@ void RightHumanWave::SetWave(RightHand* righthand) {
 	int action = righthand->GetAction();
 	XMFLOAT3 pos = righthand->GetPosition();
 	int pat = righthand->GetPat();
-	int HP = righthand->GetHP();
+	int HP = (int)righthand->GetHP();
 	//ÕŒ‚”g‚ÌˆÊ’u‚ðŒˆ‚ß‚é
 	if (action == 0 && pos.y == 1.5f && pat == 5 && !expand && HP > 0) {
 		this->pos.x = pos.x;
